+++ docs/release-notes.rst
@@ -37,6 +37,7 @@ Version 1.2.0, 13 Aug 2019
 * Update App manager client to spec version 0.1.7
 * Update E2 manager client to spec version 20190813
 * Add controller for page refresh of Angular routes
+* Display AC xAPP metrics data on dashboard
 
 Version 1.0.5, 5 July 2019
 --------------------------